// src/features/auth/schemas/login.schema.ts
import { z } from "zod";

export const loginSchema = z.object({
  email: z.email("Please enter a valid email address").trim(),

  password: z.string().min(8, "Password must be at least 8 characters"),

  captcha: z.string().min(1, "Captcha is required"),
});

export type LoginFormData = z.infer<typeof loginSchema>;